fix: changed float to decimal

+++ backend/schemas.py
@@ -1,14 +1,15 @@
 from pydantic import BaseModel, Field
+from decimal import Decimal
 
 class ProductCreate(BaseModel):
     name: str = Field(min_length=1, max_length=100)
     description: str = Field(min_length=1, max_length=100)
-    price: float = Field(gt=0)
+    price: Decimal = Field(gt=0)
 
 class ProductUpdate(BaseModel):
     name: str| None = Field(default=None,min_length=1, max_length=100)
     description: str| None = Field(default=None,min_length=1, max_length=100)
-    price: float| None = Field(default=None,gt=0)
+    price: Decimal| None = Field(default=None,gt=0)
     is_active: bool| None = None
 
 class ProductResponseImage(BaseModel):
@@ -23,7 +24,7 @@ class ProductResponse(BaseModel):
     id: int
     name: str
     description: str
-    price: float
+    price: Decimal
     is_active: bool
     images: list[ProductResponseImage] = []
